修复了部分图表数值仍为整数的bug

+++ back-end/src/main/java/com/java/pojo/Bar.java
@@ -4,13 +4,13 @@ public class Bar {
 	
 	private String name;
 	
-	private Integer value;
+	private Float value;
 
-	public Integer getValue() {
+	public Float getValue() {
 		return value;
 	}
 
-	public void setValue(Integer value) {
+	public void setValue(Float value) {
 		this.value = value;
 	}
