+++ back-end/src/main/java/com/java/pojo/CLine.java
@@ -6,7 +6,7 @@ public class CLine {
 	
 	private String name;
 	
-	private Integer value;
+	private Float value;
 
 	public String getBank() {
 		return bank;
@@ -24,14 +24,12 @@ public class CLine {
 		this.name = name;
 	}
 
-	public Integer getValue() {
+	public Float getValue() {
 		return value;
 	}
 
-	public void setValue(Integer value) {
+	public void setValue(Float value) {
 		this.value = value;
 	}
-
-
 	
 }
